runner: Set the BASH_XTRACEFD and PS4 variables explicitly
Set the BASH_XTRACEFD and PS4 variables explicitly as opposed
to setting them on the command line. If they are the first element
on the command line, argv[0] will not be "bash" and bash will
run in posix-compatible mode, which isn't always the desired
behaviour.

No tests can be provided, as the interaction between BASH_XTRACEFD
and bash-only syntax on BSD is flakey at best, which means that
tests will only pass a part of the time.

Fixes #5