Ensure that `SimpleCov.result?` returns true
at process exit.

SimpleCov runs certain finalization tasks only if `SimpleCov.result?` is
true. This includes causing the process to exit with nonzero status if
the percentage of covered lines is less than
`SimpleCov.minimum_coverage`. To trigger this task and others, this
changeset introduces a refinement on the SimpleCov singleton class that
permits storing our own `SimpleCov::Result` object as the value of
`SimpleCov.result`.

Because this changeset requires monkeypatching the SimpleCov singleton
class, it includes a warning that will be triggered if
`SimpleCov.result` is updated upstream to accept arguments (potentially
meaning that it is now an accessor method and not just a reader).