Redefine `BASH_VERSION` when `bash_path` is read from command options
This may be a little hacky, but prevents corner-cases where `BASH_VERSION` was set based on the default `/bin/bash` but the actual version being run was changed with `--bash-path`.
This is especially important on macOS, where `/bin/bash` is a very old version that doesn't support `BASH_XTRACEFD` but another more-recent version is installed.

Fixes #57